fix pfff~ replace long by int32
had to change type of buffer from ‘long’ to ‘int32’. same for mask (and
also kLongMin).

+++ LitterLib/LitterLib.h
@@ -185,12 +185,16 @@ enum CategoryClass {
 
 
 	// Some constants we need. This stuff is simple enough to handle with macros
+// vb: had to change these from long to int to make pfff~ work - ok???
 #ifndef kULongMax
 	#define kULongMax			((unsigned long) 0xffffffff)
-	#define kLongMax			((long) 0x7fffffff)
-	#define kLongMin			((long) 0x80000000)
+	//#define kLongMax			((long) 0x7fffffff)
+    #define kLongMax			((int) 0x7fffffff)
+	//#define kLongMin			((long) 0x80000000)
+    #define kLongMin			((int) 0x80000000)
 	#define kUShortMax			((unsigned short) 0xffff)
 #endif
+
 
 	// For floating point constants it's better to use extern const...
 extern const double kPi,
